pyproject.toml
[project]
name = "biobert-mlflow-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azureml-core>=1.60.0",
    "azureml-mlflow>=1.60.0.post1",
    "fastembed>=0.6.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "setuptools>=80.4.0",
]
